Rich command line history introduced. Rich history is a text file (.history-ALL) where each line contains: - date and time - username $CLE_USER-shell_pid - return code (contains '@' for shell invocation) - working directory (or terminal in case of new shell session) - command itself (or info about the source of the session)
Rich history is a separate file besides regular - per user's .history-*
files. With $CLE_USER it is trackable and searchable in multi-admin
environments. Every command is recorded immediately so the content of
.history-ALL file will contain mix of commands from multiple terminal
sessions, however they are easily distinguishable by shell_pid. This file
is not size-restricted and can keep cmdline history for ehm... years.
You might want to rotate it, but this is not covered by CLE.

There are two workaround functions:
'h' - just a plain 'history' wrapper however with nice output on
terminal
'hh' - rich history tool that allows searching for substrings. The
simple grep is involved so you can search against anything, e.g a date
or of course part of command line. Moreover, function 'hh' has options
-s shows only those commands that were successful (return code 0)
-c shows commands only without date, time, etc. And by the way, output
of 'hh' - the rich history tool is colorful :)

Built-in bash history mechanism is not altered anyhow and known Ctrl-R
searching works as expected.