This is initiation of branch 'Nova'
No new features are planned but slight (or maybe bigger) reorganization
is to be done: clerc, config and other data creates quite a mess in home
directory, scattering files here and there. This is not nice. Thus the main
goal now will be to cover the whole environment into one subfolder:

    $HOME/.cle    or rather     $HOME/.cle-username

Some exceptions might remain (e.g. rich history file $HOME/.history-ALL)
And, the result will be:
- clean homedir
- optimized routines (maybe)
- easy to maintain and backup the environment
- more easier to understand what's what

Speedup is not to be expected, but who knows and resulting filesize
will be a surprise. Also functionality will remain the same. Even if
some changes / additions might be done in modules (surely in mod-mod)

This edit will be anyway quite adventurous!
Stay tuned, watch github while 'Nova' emerges!