Even smaller history tweak.
A formatting sttring according to command `date` needs to be set to
variable HISTTIMEFORMAT to ensure the timestamps in the shell history
file. the value is set only in case the string is empty. This attempt
preserves value already set in .bashrc or .profile. In other words,
even less push and tweaks to provide maximal comfort.