Fixing Issue #75
Instead of comparing BASH_COMMAND and PROMPT_COMMAND there is a better
way: to grab the command directly from shell history and compare with
previous. This is overcomes the Midnight commander's way - adding
stuff to PROMPT_COMMAND variable. Plus it is easier to understand.
One limitation exists: if a user executes `history -n` to get the shell
history from other sessions the very last command will appear as if it
was issued in the current session.