Fixed config file behavior
Configuration may have suffix `-hostname` but plain `$CLE_D/cf` is used
when there is not the file with suffix. This is useful in case the
hostname change (on workstation). The existing configuration will be
used regardless of current hostname.

suggestion: manually rename the configuration file from `cf-hostname` to
the plain `cf`